backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
mysqlclient==2.2.6
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
dj-database-url==2.3.0
Pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
supabase-py==2.14.0
django-storages==1.14.4